test: split the tally tests off the completion
The completion module had grown to hold two subjects: what a run's shares
add up to, and which values count as a share's tally at all.  They are
tested against different things -- one against the run row, the other
against the reader that produces the counts -- and the second is where
the guards that keep a foreign line out of an account live.

A move only: the four tests that ask what a tally is are byte-identical
under their new headers, and the completion module names where they went.
